fix(refactor): reduce planner audit surface
