+++ src/core/refactor/planner.rs
@@ -9,8 +9,12 @@ use crate::undo::UndoSnapshot;
 use crate::utils::autofix::{self, FixApplied, FixResultsSummary};
 use crate::Error;
 use serde::Serialize;
-use std::collections::{BTreeMap, BTreeSet, HashSet};
+use std::collections::{BTreeSet, HashSet};
 use std::path::{Path, PathBuf};
+
+use super::sandbox::{
+    clone_tree, copy_changed_files, diff_tree_snapshots, snapshot_tree, SandboxDir,
+};
 
 pub const KNOWN_PLAN_SOURCES: &[&str] = &["audit", "lint", "test"];
 
@@ -246,7 +250,7 @@ pub fn build_refactor_plan(request: RefactorPlanRequest) -> crate::Result<Refact
     })
 }
 
-pub fn normalize_sources(sources: &[String]) -> crate::Result<Vec<String>> {
+fn normalize_sources(sources: &[String]) -> crate::Result<Vec<String>> {
     let lowered: Vec<String> = sources.iter().map(|source| source.to_lowercase()).collect();
     let unknown: Vec<String> = lowered
         .iter()
@@ -623,146 +627,6 @@ fn run_test_stage(
     })
 }
 
-struct SandboxDir {
-    path: PathBuf,
-}
-
-impl SandboxDir {
-    fn path(&self) -> &Path {
-        &self.path
-    }
-}
-
-impl Drop for SandboxDir {
-    fn drop(&mut self) {
-        let _ = std::fs::remove_dir_all(&self.path);
-    }
-}
-
-fn clone_tree(src: &Path) -> crate::Result<SandboxDir> {
-    let temp = std::env::temp_dir().join(format!("homeboy-refactor-ci-{}", uuid::Uuid::new_v4()));
-    std::fs::create_dir_all(&temp).map_err(|e| {
-        Error::internal_io(
-            e.to_string(),
-            Some("create temp refactor sandbox".to_string()),
-        )
-    })?;
-    copy_dir_recursive(src, &temp)?;
-    Ok(SandboxDir { path: temp })
-}
-
-fn copy_dir_recursive(src: &Path, dst: &Path) -> crate::Result<()> {
-    std::fs::create_dir_all(dst)
-        .map_err(|e| Error::internal_io(e.to_string(), Some("create sandbox dir".to_string())))?;
-
-    for entry in std::fs::read_dir(src)
-        .map_err(|e| Error::internal_io(e.to_string(), Some("read source dir".to_string())))?
-    {
-        let entry = entry
-            .map_err(|e| Error::internal_io(e.to_string(), Some("read dir entry".to_string())))?;
-        let src_path = entry.path();
-        let dst_path = dst.join(entry.file_name());
-
-        if src_path.is_dir() {
-            if entry.file_name() == ".git" {
-                continue;
-            }
-            copy_dir_recursive(&src_path, &dst_path)?;
-        } else {
-            std::fs::copy(&src_path, &dst_path).map_err(|e| {
-                Error::internal_io(e.to_string(), Some("copy sandbox file".to_string()))
-            })?;
-        }
-    }
-
-    Ok(())
-}
-
-fn copy_changed_files(
-    src_root: &Path,
-    dst_root: &Path,
-    changed_files: &[String],
-) -> crate::Result<()> {
-    for file in changed_files {
-        let src = src_root.join(file);
-        let dst = dst_root.join(file);
-
-        if let Some(parent) = dst.parent() {
-            std::fs::create_dir_all(parent).map_err(|e| {
-                Error::internal_io(e.to_string(), Some(format!("create parent for {}", file)))
-            })?;
-        }
-
-        std::fs::copy(&src, &dst).map_err(|e| {
-            Error::internal_io(e.to_string(), Some(format!("copy changed file {}", file)))
-        })?;
-    }
-
-    Ok(())
-}
-
-fn snapshot_tree(root: &str) -> crate::Result<BTreeMap<String, u64>> {
-    let root_path = Path::new(root);
-    let mut files = BTreeMap::new();
-    snapshot_tree_recursive(root_path, root_path, &mut files)?;
-    Ok(files)
-}
-
-fn snapshot_tree_recursive(
-    root: &Path,
-    dir: &Path,
-    files: &mut BTreeMap<String, u64>,
-) -> crate::Result<()> {
-    for entry in std::fs::read_dir(dir)
-        .map_err(|e| Error::internal_io(e.to_string(), Some("read sandbox dir".to_string())))?
-    {
-        let entry = entry.map_err(|e| {
-            Error::internal_io(e.to_string(), Some("read sandbox entry".to_string()))
-        })?;
-        let path = entry.path();
-
-        if path.is_dir() {
-            snapshot_tree_recursive(root, &path, files)?;
-            continue;
-        }
-
-        let relative = path
-            .strip_prefix(root)
-            .map_err(|e| {
-                Error::internal_io(e.to_string(), Some("strip sandbox prefix".to_string()))
-            })?
-            .to_string_lossy()
-            .replace('\\', "/");
-        let metadata = std::fs::metadata(&path).map_err(|e| {
-            Error::internal_io(e.to_string(), Some("stat sandbox file".to_string()))
-        })?;
-        files.insert(relative, metadata.len());
-    }
-
-    Ok(())
-}
-
-fn diff_tree_snapshots(
-    before: &BTreeMap<String, u64>,
-    after: &BTreeMap<String, u64>,
-) -> Vec<String> {
-    let mut changed = BTreeSet::new();
-
-    for (file, size) in after {
-        if before.get(file) != Some(size) {
-            changed.insert(file.clone());
-        }
-    }
-
-    for file in before.keys() {
-        if !after.contains_key(file) {
-            changed.insert(file.clone());
-        }
-    }
-
-    changed.into_iter().collect()
-}
-
 fn collect_audit_changed_files(fix_result: &fixer::FixResult) -> Vec<String> {
     let mut files = BTreeSet::new();
     for fix in &fix_result.fixes {
@@ -802,7 +666,7 @@ fn summarize_audit_fix_result_entries(fix_result: &fixer::FixResult) -> Vec<FixA
     entries
 }
 
-pub fn analyze_stage_overlaps(stages: &[PlanStageSummary]) -> Vec<PlanOverlap> {
+fn analyze_stage_overlaps(stages: &[PlanStageSummary]) -> Vec<PlanOverlap> {
     let mut overlaps = Vec::new();
 
     for (later_index, later_stage) in stages.iter().enumerate() {
@@ -847,10 +711,7 @@ pub fn analyze_stage_overlaps(stages: &[PlanStageSummary]) -> Vec<PlanOverlap> {
     overlaps
 }
 
-pub fn summarize_plan_totals(
-    stages: &[PlanStageSummary],
-    total_files_selected: usize,
-) -> PlanTotals {
+fn summarize_plan_totals(stages: &[PlanStageSummary], total_files_selected: usize) -> PlanTotals {
     PlanTotals {
         stages_with_proposals: stages
             .iter()
